Authentication Implemented

+++ src/main/java/org/kainos/ea/resources/EmployeeController.java
@@ -1,26 +1,35 @@
 package org.kainos.ea.resources;
 
 import io.swagger.annotations.Api;
+import org.kainos.ea.api.AuthService;
 import org.kainos.ea.api.EmployeeService;
 import org.kainos.ea.cli.DeliveryEmployeeRequest;
+import org.kainos.ea.cli.UserRole;
+import org.kainos.ea.client.AuthenticationException;
 import org.kainos.ea.client.GenericActionFailedException;
 import org.kainos.ea.client.GenericValidationException;
 import org.kainos.ea.client.GenericDoesNotExistException;
+import org.kainos.ea.db.AuthDao;
 
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-@Api("Runtime Terrors API")
+@Api("So SOLID Crew API")
 @Path("/api")
 public class EmployeeController {
     private final EmployeeService employeeService = new EmployeeService();
+    private final AuthService authService = new AuthService(new AuthDao());
 
     @GET
     @Path("/employees")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response getAllOrders() {
+    public Response getAllOrders(@QueryParam("token") String token) {
         try {
+            if(!authService.doesTokenHaveRole(token,UserRole.HR)){
+                return Response.status(Response.Status.FORBIDDEN).build();
+            }
+
             return Response
                     .status(Response.Status.OK)
                     .entity(employeeService.getAllEmployees())
@@ -28,14 +37,21 @@ public class EmployeeController {
         } catch (GenericActionFailedException e) {
             System.err.println(e.getMessage());
             return Response.serverError().build();
+        } catch (AuthenticationException e) {
+            System.err.println(e.getMessage());
+            return Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build();
         }
     }
 
     @POST
     @Path("/employees/delivery")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response createOrder(DeliveryEmployeeRequest deliveryEmployee) {
+    public Response createOrder(DeliveryEmployeeRequest deliveryEmployee, @QueryParam("token") String token) {
         try {
+            if(!authService.doesTokenHaveRole(token,UserRole.HR)){
+                return Response.status(Response.Status.FORBIDDEN).build();
+            }
+
             return Response
                     .ok(employeeService.createDeliveryEmployee(deliveryEmployee))
                     .build();
@@ -48,14 +64,21 @@ public class EmployeeController {
                     .status(Response.Status.BAD_REQUEST)
                     .entity(e.getMessage())
                     .build();
+        } catch (AuthenticationException e) {
+            System.err.println(e.getMessage());
+            return Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build();
         }
     }
 
     @GET
     @Path("/employees/delivery/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response getDeliveryEmployeeById(@PathParam("id") int id) {
+    public Response getDeliveryEmployeeById(@PathParam("id") int id,@QueryParam("token") String token) {
         try {
+            if(!authService.doesTokenHaveRole(token,UserRole.HR)){
+                return Response.status(Response.Status.FORBIDDEN).build();
+            }
+
             return Response
                     .status(Response.Status.OK)
                     .entity(employeeService.getDeliveryEmployeeById(id))
@@ -64,8 +87,11 @@ public class EmployeeController {
             System.err.println(e.getMessage());
             return Response.serverError().build();
         } catch (GenericDoesNotExistException e){
-
+            System.err.println(e.getMessage());
             return Response.status(Response.Status.NOT_FOUND).entity(e.getMessage()).build();
+        } catch (AuthenticationException e) {
+            System.err.println(e.getMessage());
+            return Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build();
         }
     }
 
@@ -73,8 +99,12 @@ public class EmployeeController {
     @GET
     @Path("/employees/delivery")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response getAllDeliveryEmployees(){
+    public Response getAllDeliveryEmployees(@QueryParam("token") String token){
         try {
+            if(!authService.doesTokenHaveRole(token,UserRole.HR)){
+                return Response.status(Response.Status.FORBIDDEN).build();
+            }
+
             return Response
                     .status(Response.Status.OK)
                     .entity(employeeService.getAllDeliveryEmployees())
@@ -82,15 +112,20 @@ public class EmployeeController {
         } catch (GenericActionFailedException e) {
             System.err.println(e.getMessage());
             return Response.serverError().build();
+        } catch (AuthenticationException e) {
+            System.err.println(e.getMessage());
+            return Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build();
         }
     }
 
     @DELETE
     @Path("/employees/delivery/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response deleteDeliveryEmployee(@PathParam("id") int id) {
-
+    public Response deleteDeliveryEmployee(@PathParam("id") int id, @QueryParam("token") String token) {
         try {
+            if(!authService.doesTokenHaveRole(token,UserRole.HR)){
+                return Response.status(Response.Status.FORBIDDEN).build();
+            }
 
             employeeService.deleteDeliveryEmployee(id);
 
@@ -98,11 +133,39 @@ public class EmployeeController {
 
         } catch (GenericActionFailedException e) {
             System.err.println(e.getMessage());
-
             return Response.serverError().build();
         } catch (GenericDoesNotExistException e) {
-
+            System.err.println(e.getMessage());
             return Response.status(Response.Status.NOT_FOUND).entity(e.getMessage()).build();
+        } catch (AuthenticationException e) {
+            System.err.println(e.getMessage());
+            return Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build();
+        }
+    }
+
+
+    @PUT
+    @Path("/employees/delivery/{id}")
+    @Produces(MediaType.APPLICATION_JSON)
+    public Response createEmployee(@PathParam("id") int id, DeliveryEmployeeRequest employee,@QueryParam("token") String token) {
+        try{
+            if(!authService.doesTokenHaveRole(token,UserRole.HR)){
+                return Response.status(Response.Status.FORBIDDEN).build();
+            }
+
+            employeeService.updateEmployee(id, employee);
+
+            return Response.ok().build();
+        } catch (GenericActionFailedException e) {
+            System.err.println(e.getMessage());
+            return Response.serverError().build();
+
+        } catch (GenericValidationException | GenericDoesNotExistException | AuthenticationException e) {
+            System.err.println(e.getMessage());
+            return Response
+                    .status(Response.Status.BAD_REQUEST)
+                    .entity(e.getMessage())
+                    .build();
         }
     }
 }
